Keep a memory of loaded versions.
Each time the lib loads, append its version onto the ended of private list
`loaded`. This tracks which revisions of the lib get loaded, in which order.
If at any point in the future, a bug occurs that impacts the ingtegrity of the
internal state, or must otherwise be fixed or accounted for, the lib can search
the loaded table to see if affected versions were present.